+++ src/authentication.ts
@@ -6,5 +6,6 @@ export {Authorize} from './common/authorize.decorator';
 export {AuthorizeDirective} from './common/authorize.directive';
 export {AccessToken} from './common/accessToken';
 export {AuthorizationModule} from './modules/authorization.module';
+export {AuthInterceptorOptions} from './authInterceptor/authInterceptor.options';
 export * from './authInterceptor/authInterceptor';
 export * from './suppressAuthInterceptor/suppressAuthInterceptor';